refactor: 게임 진행 프로세스 내 관심사끼리 하위 폴더로 이동
