+++ src/Models/RacingGame/createRacingGame.js
@@ -1,19 +1,27 @@
-import createCars from '../Cars/createCars.js';
-import RandomNumberStrategy from '../MoveStrategy/RandomNumberStrategy/RandomNumberStrategy.js';
-import validateRoundCount from '../Round/validateTotalRound.js';
+import validateCarNames, { CAR_NAMES_SEPERATOR } from './validateCarNames.js';
+import validateTotalRound from './validateTotalRound.js';
+import { splitAndTrim } from '../utils/utils.js';
+import Car from '../Car/Car.js';
 import Round from '../Round/Round.js';
+import RandomNumberStrategy from '../MoveStrategy/RandomNumberStrategy/RandomNumberStrategy.js';
 
 // TODO 항상 실행 순서가 보장되어야 하는 함수 처리 방법
 // 해결: Flag 도입
 export default function createRacingGame(movableCondition) {
     let cars = [];
-    let totalRound = 0;
     let rounds = [];
 
-    const set = (carNames, roundsInput) => {
-        cars = createCars(carNames);
-        validateRoundCount(roundsInput);
-        totalRound = Number(roundsInput);
+    const set = (carNames, totalRound) => {
+        validateCarNames(carNames);
+        validateTotalRound(totalRound);
+
+        const carNamesArr = splitAndTrim(carNames, CAR_NAMES_SEPERATOR);
+        const totalRoundNum = Number(totalRound);
+
+        cars = carNamesArr.map((carName) => Car.of(carName));
+        rounds = Array.from({ length: totalRoundNum }, (idx) =>
+            Round.of(idx + 1),
+        );
     };
 
     const play = (
@@ -21,10 +29,8 @@ export default function createRacingGame(movableCondition) {
             () => new RandomNumberStrategy(movableCondition),
         ),
     ) => {
-        Array.from({ length: totalRound }).forEach((_, idx) => {
-            const round = Round.of(idx + 1);
+        rounds.map((round) => {
             cars = round.run(cars, moveStrategies);
-            rounds.push(round);
         });
     };
