+++ src/Models/utils/utils.js
@@ -11,3 +11,6 @@ export const isEmptyValue = (target) => {
         (isString(target) && isEmptyString(target))
     );
 };
+
+export const splitAndTrim = (target, separator) =>
+    target.split(separator).map((str) => str.trim());
